*** empty log message ***
git-svn-id: http://sqlworkbench.mgm-tp.com/sqlworkbench/trunk/sqlworkbench@165 6c36a7f0-0884-4472-89c7-80b7fa7b7f28

+++ src/workbench/gui/dbobjects/SpoolerProgressPanel.java
@@ -28,9 +28,9 @@ public class SpoolerProgressPanel extends javax.swing.JPanel
 		initComponents();
 	}
 	
-	public synchronized void setRowInfo(int aRow)
+	public synchronized void setRowInfo(long aRow)
 	{
-		this.rowInfo.setText(Integer.toString(aRow));
+		this.rowInfo.setText(Long.toString(aRow));
 	}
 	
 	public synchronized void setInfoText(String aText)
